+++ packages/coding-agent/test/suite/regressions/5943-session-start-notify.test.ts
@@ -16,6 +16,8 @@ function createUiContext(
 		input: async () => undefined,
 		notify: onNotify,
 		onTerminalInput: () => () => {},
+		getTranscriptSelection: () => undefined,
+		onTranscriptSelection: () => () => {},
 		setStatus: () => {},
 		setWorkingMessage: () => {},
 		setWorkingVisible: () => {},